Fix circular dependency issue that was making Jest unhappy
